use speech synthesis for move confirmation when both are enabled
